IEP-897: refactoring and installing freertos-gdb package
