+++ bundles/com.espressif.idf.core/src/com/espressif/idf/core/tools/util/messages.properties
@@ -1,0 +1,7 @@
+InstallToolsHandler_CopyingOpenOCDRules=Copying OpenOCD Rules
+InstallToolsHandler_OpenOCDRulesCopied=Rules Copied to system
+InstallToolsHandler_OpenOCDRulesNotCopied=Rules Not Copied to system
+InstallToolsHandler_OpenOCDRulesCopyPaths=Copying File: %s to destination: %s
+InstallToolsHandler_OpenOCDRulesCopyError=Unable to copy rules for OpenOCD to system directory, try running the eclipse with sudo command
+InstallToolsHandler_OpenOCDRulesCopyWarning=Warning
+InstallToolsHandler_OpenOCDRulesCopyWarningMessage=The rules file is already present in the /etc/udev/rules.d/ directory, Do you want to replace the file with the file from OpenOCD directory
