+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/tools/Messages.java
@@ -71,13 +71,6 @@ public class Messages extends NLS
 	public static String IDFDownloadWizard_UpdatingIDFPathMessage;
 	public static String BtnCancel;
 	public static String OperationCancelledByUser;
-	public static String InstallToolsHandler_CopyingOpenOCDRules;
-	public static String InstallToolsHandler_OpenOCDRulesCopied;
-	public static String InstallToolsHandler_OpenOCDRulesCopyError;
-	public static String InstallToolsHandler_OpenOCDRulesCopyPaths;
-	public static String InstallToolsHandler_OpenOCDRulesCopyWarning;
-	public static String InstallToolsHandler_OpenOCDRulesCopyWarningMessage;
-	public static String InstallToolsHandler_OpenOCDRulesNotCopied;
 	public static String AbstractToolsHandler_ExecutingMsg;
 	public static String ToolAreadyPresent;
 	public static String ForceDownload_ToolTip;
