+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/tools/messages.properties
@@ -70,13 +70,6 @@ IDFDownloadWizard_DecompressingCompleted=Archive extracted!
 IDFDownloadWizard_UpdatingIDFPathMessage=Updating IDF_PATH to: {0}
 BtnCancel=Cancel
 OperationCancelledByUser=Operation is cancelled by the user!
-InstallToolsHandler_CopyingOpenOCDRules=Copying OpenOCD Rules
-InstallToolsHandler_OpenOCDRulesCopied=Rules Copied to system
-InstallToolsHandler_OpenOCDRulesNotCopied=Rules Not Copied to system
-InstallToolsHandler_OpenOCDRulesCopyPaths=Copying File: %s to destination: %s
-InstallToolsHandler_OpenOCDRulesCopyError=Unable to copy rules for OpenOCD to system directory, try running the eclipse with sudo command
-InstallToolsHandler_OpenOCDRulesCopyWarning=Warning
-InstallToolsHandler_OpenOCDRulesCopyWarningMessage=The rules file is already present in the /etc/udev/rules.d/ directory, Do you want to replace the file with the file from OpenOCD directory
 AbstractToolsHandler_ExecutingMsg=Executing
 ToolAreadyPresent=Tool already present in ESPRESSIF home directory.
 ForceDownloadToolTip=Downloads the tools again even if they are already downloaded in ESPRESSIF home directory.
