+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/update/Messages.java
@@ -30,6 +30,7 @@ public class Messages extends NLS
 	public static String InstallToolsHandler_ExportingPathsMsg;
 	public static String InstallToolsHandler_InstallingPythonMsg;
 	public static String InstallToolsHandler_InstallingWebscoketMsg;
+	public static String InstallToolsHandler_InstallingFreeRtosGdbtMsg;
 	public static String InstallToolsHandler_InstallingToolsMsg;
 	public static String InstallToolsHandler_ItWilltakeTimeMsg;
 	public static String InstallToolsHandler_ToolsCompleted;
