+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/update/messages.properties
@@ -23,6 +23,7 @@ InstallToolsHandler_ConfiguredCMakeMsg=Configured CMake toolchain. You can check
 InstallToolsHandler_ExportingPathsMsg=Exporting and processing tools paths
 InstallToolsHandler_InstallingPythonMsg=Installing Python
 InstallToolsHandler_InstallingWebscoketMsg=Installing websocket_client
+InstallToolsHandler_InstallingFreeRtosGdbtMsg=Installing freertos-gdb
 InstallToolsHandler_InstallingToolsMsg=Installing tools...
 InstallToolsHandler_ItWilltakeTimeMsg=This can take a while. Please be patient.
 InstallToolsHandler_ToolsCompleted=Install tools completed.
